# Merge AllReduce Worker and Reducer Kernels