--- a/tt-train/sources/ttml/autograd/graph.cpp
+++ b/tt-train/sources/ttml/autograd/graph.cpp
@@ -6,14 +6,6 @@
 
 #include <fmt/core.h>
 
-#include <chrono>
-#include <cstddef>
-#include <functional>
-#include <memory>
-#include <span>
-#include <typeinfo>
-#include <vector>
-
 #include "core/debug.hpp"
 #include "core/system_utils.hpp"