--- a/tt-train/sources/ttml/core/system_utils.cpp
+++ b/tt-train/sources/ttml/core/system_utils.cpp
@@ -6,9 +6,6 @@
 
 #include <cxxabi.h>
 
-#include <memory>
-#include <string>
-
 namespace ttml::core {
 std::string demangle(const char* name) {
     int status = -4;