--- a/tt-train/sources/ttml/datasets/in_memory_token_dataset.hpp
+++ b/tt-train/sources/ttml/datasets/in_memory_token_dataset.hpp
@@ -4,10 +4,7 @@
 
 #pragma once
 
-#include <cstdint>
 #include <span>
-#include <stdexcept>
-#include <vector>
 
 #include "dataset_base.hpp"