--- a/tt-train/sources/ttml/init/cpu_initializers.cpp
+++ b/tt-train/sources/ttml/init/cpu_initializers.cpp
@@ -4,7 +4,6 @@
 
 #include "cpu_initializers.hpp"
 
-#include <cmath>
 #include <random>
 
 #include "autograd/auto_context.hpp"
@@ -63,7 +62,7 @@ void xavier_uniform_init(std::vector<float>& vec, FanParams params) {
 
 void xavier_normal_init(std::vector<float>& vec, FanParams params) {
     auto& [fan_in, fan_out] = params;
-    float stddev = std::sqrt(2.0F / (float)(fan_in + fan_out));
+    float stddev = std::sqrtf(2.0F / (float)(fan_in + fan_out));
 
     // Random number generator with a seed
     // Mersenne Twister generator