--- a/tt-train/sources/ttml/modules/linear_module.cpp
+++ b/tt-train/sources/ttml/modules/linear_module.cpp
@@ -4,7 +4,6 @@
 
 #include "linear_module.hpp"
 
-#include <cmath>
 #include <core/ttnn_all_includes.hpp>
 
 #include "autograd/auto_context.hpp"
@@ -21,12 +20,12 @@ ttml::autograd::TensorPtr create_weight(uint32_t in_features, uint32_t out_featu
     auto* device = &autograd::ctx().get_device();
     auto weight_shape = core::create_shape({1, 1, out_features, in_features});
     auto weight = ttml::autograd::create_tensor();
-    const float init_k = std::sqrt(1.F / static_cast<float>(in_features));
+    const float init_k = std::sqrtf(1.F / static_cast<float>(in_features));
     init::uniform_init(weight, weight_shape, init::UniformRange{-init_k, init_k});
     return weight;
 }
 ttml::autograd::TensorPtr create_bias(uint32_t in_features, uint32_t out_features) {
-    const float init_k = std::sqrt(1.F / static_cast<float>(in_features));
+    const float init_k = std::sqrtf(1.F / static_cast<float>(in_features));
     auto* device = &ttml::autograd::ctx().get_device();
     auto bias_shape = ttml::core::create_shape({1, 1, 1, out_features});
     auto bias = ttml::autograd::create_tensor();