--- a/tt-train/sources/ttml/serialization/msgpack_file.hpp
+++ b/tt-train/sources/ttml/serialization/msgpack_file.hpp
@@ -9,7 +9,6 @@
 #include <span>
 #include <string>
 #include <string_view>
-#include <variant>
 #include <vector>
 
 namespace ttml::serialization {