--- a/tt-train/sources/ttml/tokenizers/bpe_tokenizer.hpp
+++ b/tt-train/sources/ttml/tokenizers/bpe_tokenizer.hpp
@@ -5,9 +5,6 @@
 #pragma once
 
 #include <cstdint>
-#include <memory>
-#include <string>
-#include <vector>
 
 #include "tokenizer_base.hpp"