--- a/tt-train/sources/ttml/tokenizers/char_tokenizer_trainer.hpp
+++ b/tt-train/sources/ttml/tokenizers/char_tokenizer_trainer.hpp
@@ -3,9 +3,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #pragma once
-#include <memory>
-#include <string>
-
 #include "char_tokenizer.hpp"
 
 namespace ttml::tokenizers {