… .github/workflows/all-post-commit-workflows.yaml
… .github/workflows/build-artifact.yaml
… .github/workflows/build-wrapper.yaml
… .github/workflows/cpp-ttnn-project.yaml
… .github/workflows/tg-unit-tests-impl.yaml
… .github/workflows/umd-unit-tests.yaml
… CMakePresets.json
… tests/ttnn/unit_tests/gtests/tensor/common_tensor_test_utils.cpp
… tests/ttnn/unit_tests/operations/test_moreh_group_norm.py
… tt-train/CMakeLists.txt
… tt-train/sources/ttml/CMakeLists.txt
… tt-train/tests/datasets/dataloader_test.cpp
… ttnn/cpp/ttnn/operations/experimental/ccl/all_reduce_async/device/kernels/dataflow/reduction_receiver.cpp
… ttnn/cpp/ttnn/operations/experimental/ccl/all_reduce_async/device/kernels/dataflow/worker_reader.cpp
… ttnn/cpp/ttnn/operations/experimental/ccl/all_reduce_async/device/kernels/dataflow/worker_writer.cpp