### Problem description
Presently, the AllReduce Async Op has restriction to place CCL workers on cores other than the reducer cores.
This PR aims to merge mainly reader kernels for worker and reducer so that the workers can be placed anywhere in the sub device coregrid.

This kernel merging task handles these cases when a core is
1. CCL worker as well as a reducer.
2. CCL worker but not a reducer.
3. Not a CCL worker but only a reducer.
4. Neither a CCL worker nor a Reducer

### Optimal Worker Placement Heuristic
1. Get top row of ethernet_cores
2. Convert them virtual positions to logical cores.
3. Place worker core just below or as close to ethernet core inside the subdevice boundary
i. Shift ethernet core X inside the subdevice boundary
ii. place first worker just below the adjusted ethernet core.
iii. if there is already a worker placed at that position, shift the worker core one position down.

### Ethernet Cores and Worker Placement
Writer NOC 0
Reader NOC 1
Note: Green circles marks ethernet core, Yellow stars mark workers
### Cluster Axis 0 (4 links)
![image (2)](https://github.com/user-attachments/assets/e455a1b5-2979-473a-a216-80f20f5319b8)
### Cluster Axis 1 (3 links)
![image (3)](https://github.com/user-attachments/assets/e166a74b-bf75-488f-8c3f-a16a36cb951b)


### What's changed
1. Program factory is updated to pass dynamic sized runtime args based on the type of kernel.
3. reduction_reciever kernel is not used anymore and merged with worker_reader kernel.
5. Update local semaphore set and reset condition to avoid race condition when a core can be worker or reducer.

### Perf before and after
| Op Instance | Time Before (us) | Time After (us) | ~Savings (us) |
|-|-|-|-|
| all-reduce ff1     | 19.18  | 19.07 | +0.11 |
| all-reduce ff2     | 18.55 | 20.10 | -1.55 |
| all-reduce qkv     | 11.9  | 12.66 | -0.76 |
| all-reduce lm_head | 61.82 | 60.79| +1.03 |

### Placement Experiment Result
* If we place the workers as originally as in `all-reduce qkv`, the merged kernels arenot able to reproduce the performance as before. In short, worker placement is ineffective in case of QKV.

| QKV Placement (sender cores)                                   | Min (µs) | Max (µs) | Avg (µs) (baseline : 10.77µs)     |
|---------------------------------------------------|----------|----------|--------------|
| {(x=2, y=0), (x=3, y=0), (x=5, y=0)}               | 9.201    | 13.335   | 11.187       |
| {(x=2, y=0), (x=3, y=1), (x=5, y=2)}               | 9.770    | 13.234   | 11.101       |
| {(x=1, y=7), (x=2, y=8), (x=3, y=9)}  | 9.389    | 12.915   | 11.029       |



### Checklist
- [ ] [All post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml) CI passes
- [ ] [Blackhole Post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml) CI passes (if applicable)
- [ ] [Model regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-models.yaml) CI passes (if applicable)
- [ ] [Device performance regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-device-models.yaml) CI passes (if applicable)
- [ ] **(For models and ops writers)** Full [new models tests](https://github.com/tenstorrent/tt-metal/actions/workflows/full-new-models-suite.yaml) CI passes (if applicable)
- [ ] New/Existing tests provide coverage for changes